fixed #125, thx

+++ libelfsh/vectors_register.c
@@ -3,7 +3,7 @@
 ** @ingroup libelfsh
 **
 ** vectors_register.c for libelfsh (The ELF shell library)
-** 
+**
 ** API for registering handlers in vectors of Libelfsh
 **
 ** $Id$$
@@ -16,18 +16,18 @@ int		elfsh_register_readmemf(u_int hostype, u_int exectype, void *fct)
   vector_t	*mem;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   if (hostype >= ELFSH_HOST_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Host type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Host type", -1);
   if (exectype >= LIBELFSH_MODE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid execution Mode", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid execution Mode", -1);
   mem = aspect_vector_get(ELFSH_HOOK_READMEMF);
   dim    = alloca(sizeof(u_int) * 4);
   dim[0] = hostype;
   dim[1] = exectype;
-  aspect_vectors_insert(mem, dim, (int) fct);
+  aspect_vectors_insert(mem, dim, (unsigned long) fct);
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
 
@@ -37,18 +37,18 @@ int		elfsh_register_writememf(u_int hostype, u_int exectype, void *fct)
   vector_t	*mem;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   if (hostype >= ELFSH_HOST_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Host type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Host type", -1);
   if (exectype >= LIBELFSH_MODE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid execution Mode", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid execution Mode", -1);
   mem = aspect_vector_get(ELFSH_HOOK_WRITEMEMF);
   dim    = alloca(sizeof(u_int) * 4);
   dim[0] = hostype;
   dim[1] = exectype;
-  aspect_vectors_insert(mem, dim, (int) fct);
+  aspect_vectors_insert(mem, dim, (unsigned long) fct);
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
 
@@ -58,18 +58,18 @@ int		elfsh_register_readmem(u_int ostype, u_int devicetype, void *fct)
   vector_t	*mem;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   if (devicetype >= ELFSH_IOTYPE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid Operating System type", -1);
   mem = aspect_vector_get(ELFSH_HOOK_READMEM);
   dim    = alloca(sizeof(u_int) * 3);
   dim[0] = ostype;
   dim[1] = devicetype;
-  aspect_vectors_insert(mem, dim, (int) fct);
+  aspect_vectors_insert(mem, dim, (unsigned long) fct);
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
 
@@ -79,46 +79,46 @@ int		elfsh_register_readmema(u_int ostype, u_int devicetype, void *fct)
   vector_t	*mem;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   if (devicetype >= ELFSH_IOTYPE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid Operating System type", -1);
   mem = aspect_vector_get(ELFSH_HOOK_READMEMA);
   dim    = alloca(sizeof(u_int) * 3);
   dim[0] = ostype;
   dim[1] = devicetype;
-  aspect_vectors_insert(mem, dim, (int) fct);
+  aspect_vectors_insert(mem, dim, (unsigned long) fct);
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
 
 
 int		elfsh_register_writemem(u_int ostype, u_int devicetype, void *fct)
-			    
+
 {
   vector_t	*mem;
   u_int		*dim;
 
   PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   if (devicetype >= ELFSH_IOTYPE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		 "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                 "Invalid Operating System type", -1);
   mem = aspect_vector_get(ELFSH_HOOK_WRITEMEM);
   dim    = alloca(sizeof(u_int) * 3);
   dim[0] = ostype;
   dim[1] = devicetype;
-  aspect_vectors_insert(mem, dim, (int) fct);  
+  aspect_vectors_insert(mem, dim, (unsigned long) fct);
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
 
 
 /**
- * Registration handlers 
+ * Registration handlers
  *
  * @param archtype
  * @param objtype
@@ -126,10 +126,10 @@ int		elfsh_register_writemem(u_int ostype, u_int devicetype, void *fct)
  * @param fct
  * @return
  */
-int	elfsh_register_altplthook(u_char archtype, 
-				  u_char objtype, 
-				  u_char ostype, 
-				  void	 *fct)
+int	elfsh_register_altplthook(u_char archtype,
+                                  u_char objtype,
+                                  u_char ostype,
+                                  void   *fct)
 {
   vector_t	*altplt;
   u_int		*dim;
@@ -138,14 +138,14 @@ int	elfsh_register_altplthook(u_char archtype,
   altplt = aspect_vector_get(ELFSH_HOOK_ALTPLT);
 
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -157,7 +157,7 @@ int	elfsh_register_altplthook(u_char archtype,
 
 
 /**
- * Register an EXTPLT handler 
+ * Register an EXTPLT handler
  *
  * @param archtype
  * @param objtype
@@ -165,26 +165,26 @@ int	elfsh_register_altplthook(u_char archtype,
  * @param fct
  * @return
  */
-int		elfsh_register_extplthook(u_char archtype, 
-					  u_char objtype, 
-					  u_char ostype, 
-					  void	 *fct)
+int		elfsh_register_extplthook(u_char archtype,
+                                          u_char objtype,
+                                          u_char ostype,
+                                          void   *fct)
 {
   vector_t	*extplt;
   u_int		*dim;
 
   PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
-  
+
   extplt = aspect_vector_get(ELFSH_HOOK_EXTPLT);
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -196,7 +196,7 @@ int		elfsh_register_extplthook(u_char archtype,
 
 
 /**
- * Register an PLT handler 
+ * Register an PLT handler
  *
  * @param archtype
  * @param objtype
@@ -204,8 +204,8 @@ int		elfsh_register_extplthook(u_char archtype,
  * @param fct
  * @return
  */
-int		elfsh_register_plthook(u_char archtype, u_char objtype, 
-				       u_char ostype, void *fct)
+int		elfsh_register_plthook(u_char archtype, u_char objtype,
+                                       u_char ostype, void *fct)
 {
   vector_t	*plt;
   u_int		*dim;
@@ -214,14 +214,14 @@ int		elfsh_register_plthook(u_char archtype, u_char objtype,
 
   plt = aspect_vector_get(ELFSH_HOOK_PLT);
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
     PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
-		      "Invalid Object type", -1);
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim    = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -231,16 +231,16 @@ int		elfsh_register_plthook(u_char archtype, u_char objtype,
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
 
-/** 
- * Register an ENCODEPLT handler 
+/**
+ * Register an ENCODEPLT handler
  * @param archtype
  * @param objtype
  * @param ostype
  * @param fct
  * @return
  */
-int	elfsh_register_encodeplthook(u_char archtype, u_char objtype, 
-				     u_char ostype, void *fct)
+int	elfsh_register_encodeplthook(u_char archtype, u_char objtype,
+                                     u_char ostype, void *fct)
 {
   vector_t	*encodeplt;
   u_int		*dim;
@@ -250,14 +250,14 @@ int	elfsh_register_encodeplthook(u_char archtype, u_char objtype,
   encodeplt = aspect_vector_get(ELFSH_HOOK_ENCODEPLT);
 
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
     PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
-		      "Invalid Object type", -1);
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -268,15 +268,15 @@ int	elfsh_register_encodeplthook(u_char archtype, u_char objtype,
 }
 
 /**
- * Register an ENCODEPLT1 handler 
+ * Register an ENCODEPLT1 handler
  * @param archtype
  * @param objtype
  * @param ostype
  * @param fct
  * @return
  */
-int	elfsh_register_encodeplt1hook(u_char archtype, u_char objtype, 
-				      u_char ostype, void *fct)
+int	elfsh_register_encodeplt1hook(u_char archtype, u_char objtype,
+                                      u_char ostype, void *fct)
 {
   vector_t	*encodeplt1;
   u_int		*dim;
@@ -286,14 +286,14 @@ int	elfsh_register_encodeplt1hook(u_char archtype, u_char objtype,
   encodeplt1 = aspect_vector_get(ELFSH_HOOK_ENCODEPLT1);
 
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
     PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
-		      "Invalid Object type", -1);
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
   dim[1] = objtype;
@@ -305,7 +305,7 @@ int	elfsh_register_encodeplt1hook(u_char archtype, u_char objtype,
 
 
 /**
- * Register an ET_REL injection handler 
+ * Register an ET_REL injection handler
  * @param archtype
  * @param objtype
  * @param ostype
@@ -313,23 +313,23 @@ int	elfsh_register_encodeplt1hook(u_char archtype, u_char objtype,
  * @return
  */
 int	elfsh_register_relhook(u_char archtype, u_char objtype, u_char ostype,
-			       void *fct)
+                               void *fct)
 {
   vector_t	*rel;
   u_int		*dim;
 
   PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   rel = aspect_vector_get(ELFSH_HOOK_REL);
-  
+
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -341,15 +341,15 @@ int	elfsh_register_relhook(u_char archtype, u_char objtype, u_char ostype,
 
 
 /**
- * Register a control flow redirection handler 
+ * Register a control flow redirection handler
  * @param archtype
  * @param objtype
  * @param ostype
  * @param fct
  * @return
  */
-int	elfsh_register_cflowhook(u_char archtype, u_char objtype, 
-				 u_char ostype, void *fct)
+int	elfsh_register_cflowhook(u_char archtype, u_char objtype,
+                                 u_char ostype, void *fct)
 {
   vector_t	*cflow;
   u_int		*dim;
@@ -358,14 +358,14 @@ int	elfsh_register_cflowhook(u_char archtype, u_char objtype,
   cflow = aspect_vector_get(ELFSH_HOOK_CFLOW);
 
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -377,31 +377,31 @@ int	elfsh_register_cflowhook(u_char archtype, u_char objtype,
 
 
 /**
- * Register a args counting redirection handler 
+ * Register a args counting redirection handler
  * @param archtype
  * @param objtype
  * @param ostype
  * @param fct
  * @return
  */
-int	elfsh_register_argchook(u_char archtype, u_char objtype, 
-				u_char ostype, void *fct)
+int	elfsh_register_argchook(u_char archtype, u_char objtype,
+                                u_char ostype, void *fct)
 {
   vector_t	*argcp;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   argcp = aspect_vector_get(ELFSH_HOOK_ARGC);
 
   if (archtype >= ELFSH_ARCH_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Architecture type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Architecture type", -1);
   if (objtype >= ELFSH_FILE_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Object type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Object type", -1);
   if (ostype >= ELFSH_OS_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Operating System type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Operating System type", -1);
 
   dim = alloca(sizeof(u_int) * 4);
   dim[0] = archtype;
@@ -424,11 +424,11 @@ int		elfsh_register_allochook(u_char hostype, void *fct)
   vector_t	*map;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   map = aspect_vector_get(ELFSH_HOOK_ALLOC);
   if (hostype >= ELFSH_HOST_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Host type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Host type", -1);
   dim = alloca(sizeof(u_int));
   dim[0] = hostype;
   aspect_vectors_insert(map, dim, (unsigned long) fct);
@@ -448,11 +448,11 @@ int		elfsh_register_mprotecthook(u_char hostype, void *fct)
   vector_t	*mprot;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   mprot = aspect_vector_get(ELFSH_HOOK_MPROTECT);
   if (hostype >= ELFSH_HOST_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Host type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Host type", -1);
   dim = alloca(sizeof(u_int));
   dim[0] = hostype;
   aspect_vectors_insert(mprot, dim, (unsigned long) fct);
@@ -471,16 +471,13 @@ int		elfsh_register_munprotecthook(u_char hostype, void *fct)
   vector_t	*munprot;
   u_int		*dim;
 
-  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);  
+  PROFILER_IN(__FILE__, __FUNCTION__, __LINE__);
   munprot = aspect_vector_get(ELFSH_HOOK_MUNPROTECT);
   if (hostype >= ELFSH_HOST_NUM)
-    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__, 
-		      "Invalid Host type", -1);
+    PROFILER_ERR(__FILE__, __FUNCTION__, __LINE__,
+                      "Invalid Host type", -1);
   dim = alloca(sizeof(u_int));
   dim[0] = hostype;
   aspect_vectors_insert(munprot, dim, (unsigned long) fct);
   PROFILER_ROUT(__FILE__, __FUNCTION__, __LINE__, 0);
 }
-
-
-
